+++ mockserver-core/src/main/java/org/mockserver/socket/tls/jdk/JDKKeyAndCertificateFactory.java
@@ -14,6 +14,7 @@ import java.security.PrivateKey;
 import java.security.cert.X509Certificate;
 import java.util.Arrays;
 
+import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.mockserver.socket.tls.jdk.CertificateSigningRequest.*;
 import static org.mockserver.socket.tls.jdk.X509Generator.*;
@@ -182,41 +183,43 @@ public class JDKKeyAndCertificateFactory implements KeyAndCertificateFactory {
 
     @Override
     public void buildAndSavePrivateKeyAndX509Certificate() {
-        try {
-            if (dynamicCertificateAuthorityUpdate()) {
-                buildAndSaveCertificateAuthorityPrivateKeyAndX509Certificate();
-            }
-            String caPrivateKey = certificateAuthorityPrivateKey();
-            X509Certificate certificateAuthorityX509Certificate = certificateAuthorityX509Certificate();
-            x509AndPrivateKey = x509Generator.generateLeafX509AndPrivateKey(
-                new CertificateSigningRequest()
-                    .setKeyPairAlgorithm(KEY_GENERATION_ALGORITHM)
-                    .setSigningAlgorithm(SIGNING_ALGORITHM)
-                    .setCommonName(ROOT_COMMON_NAME)
-                    .setCommonName(ConfigurationProperties.sslCertificateDomainName())
-                    .addSubjectAlternativeNames(ConfigurationProperties.sslSubjectAlternativeNameDomains())
-                    .addSubjectAlternativeNames(ConfigurationProperties.sslSubjectAlternativeNameIps())
-                    .setKeyPairSize(MOCK_KEY_SIZE),
-                certificateAuthorityX509Certificate.getIssuerDN().getName(),
-                caPrivateKey,
-                certificateAuthorityX509Certificate
-            );
-            if (MockServerLogger.isEnabled(DEBUG)) {
+        if (isBlank(ConfigurationProperties.privateKeyPath()) || isBlank(ConfigurationProperties.x509CertificatePath())) {
+            try {
+                if (dynamicCertificateAuthorityUpdate()) {
+                    buildAndSaveCertificateAuthorityPrivateKeyAndX509Certificate();
+                }
+                String caPrivateKey = certificateAuthorityPrivateKey();
+                X509Certificate certificateAuthorityX509Certificate = certificateAuthorityX509Certificate();
+                x509AndPrivateKey = x509Generator.generateLeafX509AndPrivateKey(
+                    new CertificateSigningRequest()
+                        .setKeyPairAlgorithm(KEY_GENERATION_ALGORITHM)
+                        .setSigningAlgorithm(SIGNING_ALGORITHM)
+                        .setCommonName(ROOT_COMMON_NAME)
+                        .setCommonName(ConfigurationProperties.sslCertificateDomainName())
+                        .addSubjectAlternativeNames(ConfigurationProperties.sslSubjectAlternativeNameDomains())
+                        .addSubjectAlternativeNames(ConfigurationProperties.sslSubjectAlternativeNameIps())
+                        .setKeyPairSize(MOCK_KEY_SIZE),
+                    certificateAuthorityX509Certificate.getIssuerDN().getName(),
+                    caPrivateKey,
+                    certificateAuthorityX509Certificate
+                );
+                if (MockServerLogger.isEnabled(DEBUG)) {
+                    mockServerLogger.logEvent(
+                        new LogEntry()
+                            .setLogLevel(DEBUG)
+                            .setMessageFormat("created new X509{}with SAN Domain Names{}and IPs{}")
+                            .setArguments(x509Certificate(), Arrays.toString(ConfigurationProperties.sslSubjectAlternativeNameDomains()), Arrays.toString(ConfigurationProperties.sslSubjectAlternativeNameIps()))
+                    );
+                }
+            } catch (Exception e) {
                 mockServerLogger.logEvent(
                     new LogEntry()
-                        .setLogLevel(DEBUG)
-                        .setMessageFormat("created new X509{}with SAN Domain Names{}and IPs{}")
-                        .setArguments(x509Certificate(), Arrays.toString(ConfigurationProperties.sslSubjectAlternativeNameDomains()), Arrays.toString(ConfigurationProperties.sslSubjectAlternativeNameIps()))
+                        .setType(LogEntry.LogMessageType.EXCEPTION)
+                        .setLogLevel(Level.ERROR)
+                        .setMessageFormat("exception while generating private key and X509 certificate")
+                        .setThrowable(e)
                 );
             }
-        } catch (Exception e) {
-            mockServerLogger.logEvent(
-                new LogEntry()
-                    .setType(LogEntry.LogMessageType.EXCEPTION)
-                    .setLogLevel(Level.ERROR)
-                    .setMessageFormat("exception while generating private key and X509 certificate")
-                    .setThrowable(e)
-            );
         }
     }
 
@@ -225,10 +228,19 @@ public class JDKKeyAndCertificateFactory implements KeyAndCertificateFactory {
     }
 
     public PrivateKey privateKey() {
-        return privateKeyFromPEM(x509AndPrivateKey.getPrivateKey());
+        if (isNotBlank(ConfigurationProperties.privateKeyPath()) && isNotBlank(ConfigurationProperties.x509CertificatePath())) {
+            return privateKeyFromPEMFile(ConfigurationProperties.privateKeyPath());
+        } else {
+            return privateKeyFromPEM(x509AndPrivateKey.getPrivateKey());
+        }
     }
 
     public X509Certificate x509Certificate() {
-        return x509FromPEM(x509AndPrivateKey.getCert());
+        if (isNotBlank(ConfigurationProperties.privateKeyPath()) && isNotBlank(ConfigurationProperties.x509CertificatePath())) {
+            return x509FromPEMFile(ConfigurationProperties.x509CertificatePath());
+        } else {
+            return x509FromPEM(x509AndPrivateKey.getCert());
+        }
     }
+
 }