+++ mockserver-netty/src/test/java/org/mockserver/netty/integration/mock/tls/inbound/ClientAuthenticationAdditionalCertificateChainMockingIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.mockserver.netty.integration.mock;
+package org.mockserver.netty.integration.mock.tls.inbound;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;